feat(rocke): add packed FP6 GEMM and mixed matrix formats

+++ dnn-providers/hip-kernel-provider/rocke/platform/tests/core/test_gfx1250_fp4_scaled_wmma.py
@@ -188,7 +188,7 @@ def test_fp4_catalog_signature_and_lowering(path, scale_type):
 @pytest.mark.parametrize(
     "changes",
     [
-        {"dtype_b": "fp8"},
+        {"dtype_b": "fp16"},
         {"matrix_path": "wmma"},
         {"K": 192},
         {"scale_dtype": "fp32"},
